implement SerializableRVRow, remove KryoSerializable from UnsafeRow
