+++ src/main/scala/is/hail/annotations/UnsafeRow.scala
@@ -78,19 +78,6 @@ class UnsafeIndexedSeq(
     output.write(a, 0, a.length)
   }
 
-  private def writeObject(out: ObjectOutputStream) {
-    out.writeObject(t)
-
-    val aos = new ByteArrayOutputStream()
-    val enc = CodecSpec.default.buildEncoder(aos)
-    enc.writeRegionValue(t, region, aoff)
-    enc.flush()
-
-    val a = aos.toByteArray
-    out.writeInt(a.length)
-    out.write(a, 0, a.length)
-  }
-
   override def read(kryo: Kryo, input: Input) {
     t = kryo.readObject(input, classOf[TArray])
 
@@ -99,20 +86,6 @@ class UnsafeIndexedSeq(
     input.readFully(a, 0, smallInOff)
     using(CodecSpec.default.buildDecoder(new ByteArrayInputStream(a))) { dec =>
 
-      region = Region()
-      aoff = dec.readRegionValue(t, region)
-
-      length = region.loadInt(aoff)
-    }
-  }
-
-  private def readObject(in: ObjectInputStream) {
-    t = in.readObject().asInstanceOf[TArray]
-
-    val smallInOff = in.readInt()
-    val a = new Array[Byte](smallInOff)
-    in.readFully(a, 0, smallInOff)
-    using(CodecSpec.default.buildDecoder(new ByteArrayInputStream(a))) { dec =>
       region = Region()
       aoff = dec.readRegionValue(t, region)
 
@@ -182,7 +155,7 @@ object UnsafeRow {
 }
 
 class UnsafeRow(var t: TBaseStruct,
-  var region: Region, var offset: Long) extends Row with KryoSerializable {
+  var region: Region, var offset: Long) extends Row {
 
   def this(t: TBaseStruct, rv: RegionValue) = this(t, rv.region, rv.offset)
 
@@ -247,58 +220,6 @@ class UnsafeRow(var t: TBaseStruct,
     if (i < 0 || i >= t.size)
       throw new IndexOutOfBoundsException(i.toString)
     !t.isFieldDefined(region, offset, i)
-  }
-
-  override def write(kryo: Kryo, output: Output) {
-    output.writeBoolean(t.isInstanceOf[TStruct])
-    kryo.writeObject(output, t)
-
-    val aos = new ByteArrayOutputStream()
-    val enc = CodecSpec.default.buildEncoder(aos)
-    enc.writeRegionValue(t, region, offset)
-    enc.flush()
-
-    val a = aos.toByteArray
-    output.writeInt(a.length)
-    output.write(a, 0, a.length)
-  }
-
-  private def writeObject(out: ObjectOutputStream) {
-    out.writeObject(t)
-
-    val aos = new ByteArrayOutputStream()
-    val enc = CodecSpec.default.buildEncoder(aos)
-    enc.writeRegionValue(t, region, offset)
-    enc.flush()
-
-    val a = aos.toByteArray
-    out.writeInt(a.length)
-    out.write(a, 0, a.length)
-  }
-
-  override def read(kryo: Kryo, input: Input) {
-    val isStruct = input.readBoolean()
-    t = kryo.readObject(input, if (isStruct) classOf[TStruct] else classOf[TTuple])
-
-    val smallInOff = input.readInt()
-    val a = new Array[Byte](smallInOff)
-    input.readFully(a, 0, smallInOff)
-    using(CodecSpec.default.buildDecoder(new ByteArrayInputStream(a))) { dec =>
-      region = Region()
-      offset = dec.readRegionValue(t, region)
-    }
-  }
-
-  private def readObject(in: ObjectInputStream) {
-    t = in.readObject().asInstanceOf[TBaseStruct]
-
-    val smallInOff = in.readInt()
-    val a = new Array[Byte](smallInOff)
-    in.readFully(a, 0, smallInOff)
-    using(CodecSpec.default.buildDecoder(new ByteArrayInputStream(a))) { dec =>
-      region = Region()
-      offset = dec.readRegionValue(t, region)
-    }
   }
 }
